+++ ios/Runner/AppDelegate.swift
@@ -1,17 +1,46 @@
 import UIKit
 import Flutter
 
+
 @UIApplicationMain
 @objc class AppDelegate: FlutterAppDelegate {
-  override func application(
-    _ application: UIApplication,
-    didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?
-  ) -> Bool {
-    if #available(iOS 10.0, *) {
-      UNUserNotificationCenter.current().delegate = self as UNUserNotificationCenterDelegate
+    override func application(
+        _ application: UIApplication,
+        didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?
+    ) -> Bool {
+        if #available(iOS 10.0, *) {
+            UNUserNotificationCenter.current().delegate = self as UNUserNotificationCenterDelegate
+        }
+        
+        let controller : FlutterViewController = window?.rootViewController as! FlutterViewController
+        
+        let transcriptionChannel = FlutterMethodChannel(
+            name: "lotti/transcribe",
+            binaryMessenger: controller.binaryMessenger)
+        
+        transcriptionChannel.setMethodCallHandler { (call, result) in
+            switch call.method {
+            case "transcribe":
+                guard let args = call.arguments as? [String: Any] else { return }
+                let audioFilePath = args["audioFilePath"] as! String
+                let modelPath = args["modelPath"] as! String
+                Task {
+                    await transcribe(audioFilePath: audioFilePath, modelPath: modelPath, result: result)
+                }
+            case "detectLanguage":
+                guard let args = call.arguments as? [String: Any] else { return }
+                let audioFilePath = args["audioFilePath"] as! String
+                let modelPath = args["modelPath"] as! String
+                Task {
+                    await detectLanguage(audioFilePath: audioFilePath, modelPath: modelPath, result: result)
+                }
+            default:
+                result(FlutterMethodNotImplemented)
+            }
+        }
+        
+        GeneratedPluginRegistrant.register(with: self)
+        return super.application(application, didFinishLaunchingWithOptions: launchOptions)
     }
-
-    GeneratedPluginRegistrant.register(with: self)
-    return super.application(application, didFinishLaunchingWithOptions: launchOptions)
-  }
 }
+