+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed:
 - Only show speech settings on macOS and iOS for now
 - Hide large whisper models
+- Upgraded dependencies
+- Preparation for Flutter 3.10
 
 ### Fixed:
 - Model download on iOS